// src/API.ts
export type Like = {
  __typename: "Like",
  id: string,
  userID: string,
  postID: string,
  User?: User | null,
  Post?: Post | null,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
  owner?: string | null,
};

export type User = {
  __typename: "User",
  id: string,
  name: string,
  email?: string | null,
  username?: string | null,
  bio?: string | null,
  website?: string | null,
  nofPosts: number,
  nofFollowers: number,
  nofFollowings: number,
  image?: string | null,
  Posts?: ModelPostConnection | null,
  Comments?: ModelCommentConnection | null,
  Likes?: ModelLikeConnection | null,
  Followers?: ModelUserFollowConnection | null,
  Followings?: ModelUserFollowConnection | null,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
  owner?: string | null,
};

export type ModelPostConnection = {
  __typename: "ModelPostConnection",
  items:  Array<Post | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};

export type Post = {
  __typename: "Post",
  id: string,
  createdAt: string,
  type: string,
  description?: string | null,
  location?: string | null,
  image?: string | null,
  images?: Array< string > | null,
  video?: string | null,
  nofComments: number,
  nofLikes: number,
  userID: string,
  User?: User | null,
  Likes?: ModelLikeConnection | null,
  Comments?: ModelCommentConnection | null,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
  owner?: string | null,
};

export type ModelLikeConnection = {
  __typename: "ModelLikeConnection",
  items:  Array<Like | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};

export type ModelCommentConnection = {
  __typename: "ModelCommentConnection",
  items:  Array<Comment | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};

export type Comment = {
  __typename: "Comment",
  id: string,
  createdAt: string,
  comment: string,
  userID: string,
  postID: string,
  User?: User | null,
  Post?: Post | null,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
  owner?: string | null,
};

export type ModelUserFollowConnection = {
  __typename: "ModelUserFollowConnection",
  items:  Array<UserFollow | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};